requirements.txt
asgiref==3.6.0
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==4.2.1
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
drf-yasg==1.21.5
gunicorn==20.1.0
httpie==3.2.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
packaging==23.1
pycparser==2.21
Pygments==2.15.1
pyrsistent==0.19.3
PySocks==1.7.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
